src/main/java/org/pcs/codingsolutions/util/dp/TribonacciNumberCalculator.java
package org.pcs.codingsolutions.util.dp;

import lombok.experimental.UtilityClass;

/**
 * Solution for:
 * <p>
 * The Tribonacci sequence Tn is defined as follows:
 * T0 = 0, T1 = 1, T2 = 1, and Tn+3 = Tn + Tn+1 + Tn+2 for n >= 0.
 * Given n, return the value of Tn.
 * <p>
 * Constraints:
 * <p>
 * 0 <= n <= 37
 * The answer is guaranteed to fit within a 32-bit integer, ie. answer <= 2^31 - 1.
 */
@UtilityClass
public class TribonacciNumberCalculator {
    public int getTribonacciNumber(int n) {
        if (n == 0) {
            return 0;
        } else if (n < 3) {
            return 1;
        }

        int t0 = 0;
        int t1 = 1;
        int t2 = 1;
        int tn = 0;
        for (int i = 3; i <= n; i++) {
            tn = t0 + t1 + t2;
            t0 = t1;
            t1 = t2;
            t2 = tn;
        }

        return tn;
    }
}
